requirements.txt
fastapi[standard]
uvicorn
langchain
langchain-community
langchain-chroma
langchain-openai
langgraph
openai
chromadb
pydantic
python-dotenv
PyPDF
docx2txt
